pgo-next-gen/gatherer/requirements.txt
psycopg2
pyyaml
arrow
